Move Aliquot::Payment to separate file
